refactor(backload-matching): extract helpers + AssignmentList + DecisionsAudit
